# src/main/java/co/moveapps/spa/core/controller/api/TaskApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.1.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package co.moveapps.spa.core.controller.api;

import co.moveapps.spa.core.controller.model.ModelApiResponse;
import co.moveapps.spa.core.controller.model.TaskRequest;
import co.moveapps.spa.core.controller.model.TaskResponse;
import co.moveapps.spa.core.exception.BusinessException;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;

import javax.validation.Valid;
import java.util.List;
import java.util.Optional;
import java.util.UUID;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2024-04-05T04:13:30.339270-05:00[America/Bogota]")
@Validated
@Controller
@Tag(name = "task", description = "Operations about task")
public interface TaskApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /task : Create a task in a system.
     * Register a new task for SPA
     *
     * @param taskRequest  (required)
     * @return success (status code 200)
     *         or unexpected error (status code 200)
     */
    @Operation(
            operationId = "createTask",
            summary = "Create a task in a system.",
            description = "Register a new task for SPA",
            tags = { "task" },
            responses = {
                    @ApiResponse(responseCode = "200", description = "success", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = TaskResponse.class))
                    }),
                    @ApiResponse(responseCode = "default", description = "unexpected error", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = ModelApiResponse.class))
                    })
            },
            security = {
                    @SecurityRequirement(name = "bearerAuth")
            }
    )
    @RequestMapping(
            method = RequestMethod.POST,
            value = "/task",
            produces = { "application/json" },
            consumes = { "application/json" }
    )

    default ResponseEntity<TaskResponse> createTask(
            @Parameter(name = "TaskRequest", description = "", required = true) @Valid @RequestBody TaskRequest taskRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"updated_at\" : \"2000-01-23T04:56:07.000+00:00\", \"description\" : \"description\", \"created_at\" : \"2000-01-23T04:56:07.000+00:00\", \"id\" : 0, \"title\" : \"title\", \"user\" : { \"firstName\" : \"firstName\", \"lastName\" : \"lastName\", \"updated_at\" : \"2000-01-23T04:56:07.000+00:00\", \"phone\" : \"phone\", \"enable\" : true, \"created_at\" : \"2000-01-23T04:56:07.000+00:00\", \"id\" : 0, \"email\" : \"email\" }, \"status\" : { \"enable\" : true, \"name\" : \"name\", \"id\" : 6 } }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * DELETE /task/{id} : Remove a task of the system.
     * Delete a existent task for SPA
     *
     * @param id Unique identifier of task (required)
     * @return success (status code 200)
     *         or unexpected error (status code 200)
     */
    @Operation(
            operationId = "deleteTask",
            summary = "Remove a task of the system.",
            description = "Delete a existent task for SPA",
            tags = { "task" },
            responses = {
                    @ApiResponse(responseCode = "200", description = "success", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = Boolean.class))
                    }),
                    @ApiResponse(responseCode = "default", description = "unexpected error", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = ModelApiResponse.class))
                    })
            },
            security = {
                    @SecurityRequirement(name = "bearerAuth")
            }
    )
    @RequestMapping(
            method = RequestMethod.DELETE,
            value = "/task/{id}",
            produces = { "application/json" }
    )

    default ResponseEntity<Boolean> deleteTask(
            @Parameter(name = "id", description = "Unique identifier of task", required = true, in = ParameterIn.PATH) @PathVariable("id") UUID id
    ) throws BusinessException {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /task : Search tasks in a system with parameters
     * Search tasks for SPA with parameters
     *
     * @param page Number of page to return (optional, default to 1)
     * @param limit Maximum number of results to return (optional, default to 100)
     * @return success (status code 200)
     *         or unexpected error (status code 200)
     */
    @Operation(
            operationId = "getAllTasks",
            summary = "Search tasks in a system with parameters",
            description = "Search tasks for SPA with parameters",
            tags = { "task" },
            responses = {
                    @ApiResponse(responseCode = "200", description = "success", content = {
                            @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = TaskResponse.class)))
                    }),
                    @ApiResponse(responseCode = "default", description = "unexpected error", content = {
                            @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = ModelApiResponse.class)))
                    })
            },
            security = {
                    @SecurityRequirement(name = "bearerAuth")
            }
    )
    @RequestMapping(
            method = RequestMethod.GET,
            value = "/task",
            produces = { "application/json" }
    )

    default ResponseEntity<List<TaskResponse>> getAllTasks(
            @Parameter(name = "page", description = "Number of page to return", in = ParameterIn.QUERY) @Valid @RequestParam(value = "page", required = false, defaultValue = "0") Integer page,
            @Parameter(name = "limit", description = "Maximum number of results to return", in = ParameterIn.QUERY) @Valid @RequestParam(value = "limit", required = false, defaultValue = "100") Integer limit
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"updated_at\" : \"2000-01-23T04:56:07.000+00:00\", \"description\" : \"description\", \"created_at\" : \"2000-01-23T04:56:07.000+00:00\", \"id\" : 0, \"title\" : \"title\", \"user\" : { \"firstName\" : \"firstName\", \"lastName\" : \"lastName\", \"updated_at\" : \"2000-01-23T04:56:07.000+00:00\", \"phone\" : \"phone\", \"enable\" : true, \"created_at\" : \"2000-01-23T04:56:07.000+00:00\", \"id\" : 0, \"email\" : \"email\" }, \"status\" : { \"enable\" : true, \"name\" : \"name\", \"id\" : 6 } }, { \"updated_at\" : \"2000-01-23T04:56:07.000+00:00\", \"description\" : \"description\", \"created_at\" : \"2000-01-23T04:56:07.000+00:00\", \"id\" : 0, \"title\" : \"title\", \"user\" : { \"firstName\" : \"firstName\", \"lastName\" : \"lastName\", \"updated_at\" : \"2000-01-23T04:56:07.000+00:00\", \"phone\" : \"phone\", \"enable\" : true, \"created_at\" : \"2000-01-23T04:56:07.000+00:00\", \"id\" : 0, \"email\" : \"email\" }, \"status\" : { \"enable\" : true, \"name\" : \"name\", \"id\" : 6 } } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /task/{id} : Search task by id
     * Search task by id for SPA
     *
     * @param id Unique identifier of task (required)
     * @return success (status code 200)
     *         or unexpected error (status code 200)
     */
    @Operation(
            operationId = "getTaskById",
            summary = "Search task by id",
            description = "Search task by id for SPA",
            tags = { "task" },
            responses = {
                    @ApiResponse(responseCode = "200", description = "success", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = TaskResponse.class))
                    }),
                    @ApiResponse(responseCode = "default", description = "unexpected error", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = ModelApiResponse.class))
                    })
            },
            security = {
                    @SecurityRequirement(name = "bearerAuth")
            }
    )
    @RequestMapping(
            method = RequestMethod.GET,
            value = "/task/{id}",
            produces = { "application/json" }
    )

    default ResponseEntity<TaskResponse> getTaskById(
            @Parameter(name = "id", description = "Unique identifier of task", required = true, in = ParameterIn.PATH) @PathVariable("id") UUID id
    ) throws BusinessException {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"updated_at\" : \"2000-01-23T04:56:07.000+00:00\", \"description\" : \"description\", \"created_at\" : \"2000-01-23T04:56:07.000+00:00\", \"id\" : 0, \"title\" : \"title\", \"user\" : { \"firstName\" : \"firstName\", \"lastName\" : \"lastName\", \"updated_at\" : \"2000-01-23T04:56:07.000+00:00\", \"phone\" : \"phone\", \"enable\" : true, \"created_at\" : \"2000-01-23T04:56:07.000+00:00\", \"id\" : 0, \"email\" : \"email\" }, \"status\" : { \"enable\" : true, \"name\" : \"name\", \"id\" : 6 } }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * PUT /task/{id} : Update a task in a system.
     * Modify fields for a existent task for SPA
     *
     * @param id Unique identifier of task (required)
     * @param taskRequest  (optional)
     * @return success (status code 200)
     *         or unexpected error (status code 200)
     */
    @Operation(
            operationId = "updateTask",
            summary = "Update a task in a system.",
            description = "Modify fields for a existent task for SPA",
            tags = { "task" },
            responses = {
                    @ApiResponse(responseCode = "200", description = "success", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = TaskResponse.class))
                    }),
                    @ApiResponse(responseCode = "default", description = "unexpected error", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = ModelApiResponse.class))
                    })
            },
            security = {
                    @SecurityRequirement(name = "bearerAuth")
            }
    )
    @RequestMapping(
            method = RequestMethod.PUT,
            value = "/task/{id}",
            produces = { "application/json" },
            consumes = { "application/json" }
    )

    default ResponseEntity<TaskResponse> updateTask(
            @Parameter(name = "id", description = "Unique identifier of task", required = true, in = ParameterIn.PATH) @PathVariable("id") UUID id,
            @Parameter(name = "TaskRequest", description = "") @Valid @RequestBody(required = false) TaskRequest taskRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"updated_at\" : \"2000-01-23T04:56:07.000+00:00\", \"description\" : \"description\", \"created_at\" : \"2000-01-23T04:56:07.000+00:00\", \"id\" : 0, \"title\" : \"title\", \"user\" : { \"firstName\" : \"firstName\", \"lastName\" : \"lastName\", \"updated_at\" : \"2000-01-23T04:56:07.000+00:00\", \"phone\" : \"phone\", \"enable\" : true, \"created_at\" : \"2000-01-23T04:56:07.000+00:00\", \"id\" : 0, \"email\" : \"email\" }, \"status\" : { \"enable\" : true, \"name\" : \"name\", \"id\" : 6 } }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}